Add `specimin-evaluation` tests to Windows CI (#324)
This PR adds 4/20 of the `specimin-evaluation` test cases to the Windows
CI to ensure that `main.py` successfully executes on various types of
test cases on Windows.
- cf-1291: test typical integration test
- cf-577: download + unzip + rename/file access errors
- cf-691: download + unzip 2
- cf-4614: slightly different than expected CF outputs

This also fixes a silent error in `typecheck_test_outputs.bat` from #319
where not all test cases were being checked. The reason behind this was
how I was handling `continue` in batch; I used `goto` within a `for`
loop, which I did not know broke the script out of the loop context
(hence stopping at the test case `methodref2`).

The CI should pass once the companion PR
njit-jerse/specimin-evaluation#3 is approved; right now we should expect
at least a few of these tasks to fail.