Preserve `MustImplementMethodsVisitor` method type ancestors (#329)
This PR addresses an issue I found in [NullAway issue
#103](https://github.com/uber/NullAway/issues/103), where compile-time
errors were caused by missing methods and types. The missing methods
were caused by the following:

**1) Methods not preserved**

Foo extends (abstract) Bar extends [Some JDK interface]

If Bar contains an abstract method also present in the JDK interface,
those implementations in Foo were removed while the abstract definition
was maintained in Bar, causing compile-time errors. A check was added to
ensure that all super abstract methods did not contain anything that
needed to be preserved (i.e. when in Foo, instead of just checking Bar,
we check Bar and the JDK interface).

**2. Classes not preserved**

If a non-target class needs its method implementations to be preserved,
and some of its return/parameter types are solvable but were untracked
so far by `usedTypeElements`, their ancestors were not preserved. For
example, if a method visited by `MustImplementMethodsVisitor` is of type
`PeekingIterator<E>` which extends `Iterator<E>`, the `extends
Iterator<E>` portion is kept in the original definition but its import
statement was removed, leading to compile-time errors. The new approach
to this is, if a type element was not previously in `usedTypeElements`,
all of its ancestors (interfaces and super classes) would also be marked
for preservation.

A test case can be found in `Issue103Test`.

Thanks!