include `java/` files in output if they are targeted (#344)
In na-97, where we target a `java/` file, the targeted file
(`java/util/atomic/concurrent/Striped64.java`, I believe) was being
excluded from the output because of these lines in `SpeciminRunner`:

```java
if (target.getKey().startsWith("java/")) {
    continue;
}
```

The reason it worked on Windows was probably because Windows uses
backslashes, and it didn't work on Unix because of the forward slash
comparison here. I changed it so that the target file is preserved in
the output no matter what, even if it starts with `java/` (or `java\`).
This should make na-97 work on both OSes.

I didn't include a test case for this, since NullAway bugs are now in
the evaluation script, but if you'd like, I could add one. Thanks!