Add guardrails to prevent and correct wrong output from JavaTypeCorrect (#339)
~~Incorporates #338, so don't review until that is merged.~~

~~I'm not sure if we want to merge this. It avoids some bad behavior on
the NA-176 test, and we could probably expand it in the future for
other, similar kinds of problems. However, it definitely does not
address the root cause: why is Specimin putting itself in a situation
where javac insists that a synthetic class must be compatible with
String?~~

~~I'm going to continue my investigation, but I'm leaving this PR here
because I may want to merge it later if the fix for the root cause is
too complex and/or impracticle.~~

This PR makes a series of changes to how our javac type correction
algorithm (in the `JavaTypeCorrect` class) handles complex cases. That
algorithm is fundamentally imprecise, so the changes in this PR should
be thought of as incremental improvements to the existing design that
avoid some of its worst pathologies. In particular, this PR:
* prevents JavaTypeCorrect from ever attempting to make a class extend
`java.lang.String`, which is final;
* makes it generally more difficult for the `SyntheticUnconstrainedType`
type variable to propagate beyond the return type to which it is
applied; and,
* replaces parameters whose types are synthetic return types that get
replaced by `SyntheticUnconstrainedType` with `java.lang.Object`.
Previously, these types would hang around and make the output
non-compilable, since they'd been marked as unused.

Together, these changes make NullAway-176 reproducible, based on my
local testing.