instead of a percentage of successes, CI reads a diff (#406)
this should help with #396. CI will do a diff over the failures in
minimization and compilation, and output results of that. This solution
still requires that someone edit files that CI references when a PR is
merged, but is less brittle. Also, this will give ASHE the ability to
fail earlier if a test fails that isn't in the saved list.